refactor: ui 를 패키지로 묶는다
