+++ app/src/main/java/nextstep/shoppingcart/ui/productdetail/ProductDetailScreen.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.productdetail
+package nextstep.shoppingcart.ui.productdetail
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
