+++ app/src/main/java/nextstep/shoppingcart/ui/products/ProductsScreen.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.products
+package nextstep.shoppingcart.ui.products
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
